+++ visualization/app/codeCharta/ui/attributeSideBar/attributeSideBarHeaderSection/nodePath/nodePath.component.spec.ts
@@ -54,7 +54,8 @@ describe("nodePathComponent", () => {
 			attributes: { unary: 2 },
 			fileCount: {
 				added: 1,
-				removed: 2
+				removed: 2,
+				changed: 3
 			}
 		}
 		isDeltaStateSelectorMock.mockImplementationOnce(() => true)
@@ -62,6 +63,41 @@ describe("nodePathComponent", () => {
 
 		const { container } = await render(NodePathComponent, { componentProperties: { node } })
 
-		expect(container.textContent.replace(/\s+/g, " ")).toContain("some/folder ( 2 files | Δ1 | Δ-2)")
+		expect(container.textContent.replace(/\s+/g, " ")).toContain("some/folder ( 2 files | Δ1 | Δ-2 | Δ3)")
+	})
+	it("should display amount of files with correct english grammar, when an empty folder is selected and delta mode is enabled", async () => {
+		const node = {
+			children: [{}] as CodeMapNode[],
+			path: "some/emptyFolder",
+			attributes: { unary: 0 },
+			fileCount: {
+				added: 0,
+				removed: 2,
+				changed: 0
+			}
+		}
+		isDeltaStateSelectorMock.mockImplementationOnce(() => true)
+		selectedNodeSelectorMock.mockImplementation(() => node)
+
+		const { container } = await render(NodePathComponent, { componentProperties: { node } })
+
+		expect(container.textContent).toContain(" 0 files ") // because "zero file" is not grammatically correct
+	})
+	it("should display amount of files with correct english grammar, when a folder with 1 file is selected and delta mode is enabled", async () => {
+		const node = {
+			children: [{}] as CodeMapNode[],
+			path: "some/folderWithOneFile",
+			attributes: { unary: 1 },
+			fileCount: {
+				added: 1,
+				removed: 0
+			}
+		}
+		isDeltaStateSelectorMock.mockImplementationOnce(() => true)
+		selectedNodeSelectorMock.mockImplementation(() => node)
+
+		const { container } = await render(NodePathComponent, { componentProperties: { node } })
+
+		expect(container.textContent).toContain(" 1 file ") // because "one files" is not grammatically correct
 	})
 })